+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.0.0] - 2019-11-20
+
 ### Fixed
 
 * Fix misleading error message when the XML sitemap cannot be created for reasons other than a non-writable parent directory
@@ -9,7 +11,7 @@
 
 ### Changed
 
-* Do not create the XML sitemap for ajax requests. 
+* Do not create the XML sitemap for ajax requests.
 
 ## [0.9.0] - 2019-10-02
 
@@ -114,7 +116,8 @@ allow installations via Composer
 
 * Initial release of the module 🐣
 
-[Unreleased]: https://github.com/wanze/SeoMaestro/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/wanze/SeoMaestro/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/wanze/SeoMaestro/releases/tag/v1.0.0
 [0.9.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.9.0
 [0.8.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.8.0
 [0.7.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.7.0